clean up users/roles endpoints
